[wip] split out elastic request builder entirely
