+++ api/api/src/main/scala/uk/ac/wellcome/platform/api/services/ImagesRequestBuilder.scala
@@ -1,0 +1,16 @@
+package uk.ac.wellcome.platform.api.services
+
+import com.sksamuel.elastic4s.ElasticDsl._
+import com.sksamuel.elastic4s._
+import com.sksamuel.elastic4s.requests.searches._
+import com.sksamuel.elastic4s.requests.searches.sort._
+
+object ImagesRequestBuilder extends ElasticsearchRequestBuilder {
+
+  val idSort: FieldSort = fieldSort("id.canonicalId").order(SortOrder.ASC)
+
+  def request(queryOptions: ElasticsearchQueryOptions,
+              index: Index,
+              scored: Boolean): SearchRequest = ???
+
+}
